requirements.txt
Django>=6.0
gunicorn>=23.0
psycopg2-binary>=2.9
Pillow>=11.0
